+++ src/main/java/org/cojen/maker/Maker.java
@@ -87,4 +87,14 @@ public interface Maker {
      * the first element can be interpreted as such.
      */
     void addAttribute(String name, Object value);
+
+    /**
+     * Defines a signature for this member, which is a string for supporting generic types.
+     * The components can be strings or types (class, ClassMaker, etc), which are concatenated
+     * into a single string. Consult the JVM specification for the signature syntax.
+     *
+     * @throws IllegalArgumentException if given an unsupported component
+     * @return this
+     */
+    Maker signature(Object... components);
 }